Shared Option definition between bsserver and server demo
